requirements.txt
langchain
langchain-core
langchain-text-splitters
langchain-groq
langgraph
chromadb
streamlit
pypdf
packaging